+++ TombEngine/Game/collision/Point.cpp
@@ -408,7 +408,7 @@ namespace TEN::Collision::Point
 		short slopeAngle = Geometry::GetSurfaceSlopeAngle(GetFloorNormal());
 		short steepSlopeAngle = (GetFloorBridgeItemNumber() != NO_VALUE) ?
 			DEFAULT_STEEP_FLOOR_SLOPE_ANGLE :
-			GetBottomSector().GetSurfaceIllegalSlopeAngle(_position.x, _position.z, true);
+			GetBottomSector().GetSurfaceSteepSlopeAngle(_position.x, _position.z, true);
 		
 		return (abs(slopeAngle) >= steepSlopeAngle);
 	}
@@ -418,7 +418,7 @@ namespace TEN::Collision::Point
 		short slopeAngle = Geometry::GetSurfaceSlopeAngle(GetCeilingNormal(), -Vector3::UnitY);
 		short steepSlopeAngle = (GetCeilingBridgeItemNumber() != NO_VALUE) ?
 			DEFAULT_STEEP_CEILING_SLOPE_ANGLE :
-			GetTopSector().GetSurfaceIllegalSlopeAngle(_position.x, _position.z, false);
+			GetTopSector().GetSurfaceSteepSlopeAngle(_position.x, _position.z, false);
 		
 		return (abs(slopeAngle) >= steepSlopeAngle);
 	}