+++ TombEngine/Specific/level.h
@@ -144,7 +144,8 @@ struct LEVEL
 extern const std::vector<GAME_OBJECT_ID> BRIDGE_OBJECT_IDS;
 
 extern std::vector<int> MoveablesIds;
-extern std::vector<int> StaticObjectsIds;
+extern int* StaticObjectsLUT;
+extern int StaticObjectsLUTSize;
 extern std::vector<int> SpriteSequencesIds;
 extern LEVEL g_Level;
 extern int SystemNameHash;