+++ CHANGELOG.md
@@ -9,8 +9,10 @@ TombEngine releases are located in this repository (alongside with Tomb Editor):
 * Fixed engine performance around bridges.
 * Fixed engine performance if weather or bubble effects are active.
 * Fixed silent crashes if loaded level is corrupted or in incorrect format.
+* Fixed occasional crashes if there are static meshes placed within room border walls.
 * Fixed incorrect clipping of scaled off-centered static meshes.
-* Fixed incorrect collision detection for off-centered meshes.
+* Fixed incorrect collision detection for off-centered moveables.
+* Fixed stutter during jumps between cameras in a flyby sequence.
 * Fixed uzi targeting issues after using flycheat.
 * Fixed snow particles not always melting on the ground.
 * Fixed enemy pickups dropping on death sectors.
@@ -33,9 +35,12 @@ TombEngine releases are located in this repository (alongside with Tomb Editor):
 * Added ricochet sounds and make the effect more prominent.
 * Allow to run the engine without title level.
 * Allow more than 1024 objects in a level.
+* Allow more than 1000 static mesh slots in a level.
 
 ### Lua API changes
+* Added Moveable:GetCollidable() and Moveable:SetCollidable() functions.
 * Added Flow.GetNextLevel() function to get script entry for incoming level, if it's about to start.
+* Added Effects.GetWind() function to get current wind speed vector.
 * Added support for transparency value in DisplayString class.
 * Added extra argument for SetAmbientTrack() function to specify if new ambient track should play from the beginning.
 * Use load camera instead of load screen by playing fixed camera from OnEnd() event and removing loadScreenFile field from level's gameflow entry.
@@ -117,8 +122,8 @@ TombEngine releases are located in this repository (alongside with Tomb Editor):
 * Added Flow.LensFlare() and Flow.Starfield() classes.
 * Added Inventory.GetUsedItem(), Inventory.SetUsedItem() and Inventory.ClearUsedItem() functions.
 * Added Input.KeyClearAll() function.
-* Added Moveable.GetJointRotation() and optional 'offset' parameter for Moveable.GetJointPosition().
-* Added Moveable.GetTargetState() function.
+* Added Moveable:GetJointRotation() and optional 'offset' parameter for Moveable.GetJointPosition().
+* Added Moveable:GetTargetState() function.
 * Added Room:GetRoomNumber() function.
 * Removed anims.monkeyAutoJump. It is now a player menu configuration.
 * Fixed Volume:GetActive() method.